jwt-authentication-api/requirements.txt
fastapi==0.135.3
pydantic-settings==2.13.1
python-multipart==0.0.22
pwdlib[argon2]==0.3.0
PyJWT==2.12.1
httpx==0.28.1
pytest==9.0.2
uvicorn==0.42.0
